A wrapper function that allows you to run simulations of a network under many different conditions.

This function allows you to set up a variety of genotype and starting node conditions to be simulated. It also allows you to test the outcome of simulations under randomly assigned starting values. The function will make sure to include one wild type condition (all genotypes set to 1) regardless of if specified by the user.

Usage

setupSims(
  folder,
  delay = 2,
  maxStep = 100,
  steadyThreshold = 4,
  exogenousSupply = FALSE,
  priorScreen = FALSE,
  robustnessTest = FALSE,
  genotypeBaseline = FALSE,
  nodestartBaseline = FALSE,
  necStimThreshold = NULL,
  altTopologyName = NULL,
  saveOutput = TRUE,
  combinatorial = TRUE,
  preventDrop = FALSE,
  reduceSize = FALSE,
  nCores = 1
)

Arguments

folder

the model folder generated by the buildModel function.

delay

the amount of delay for any delayed transport. Default set to 2.

maxStep

the maximum number of timesteps for which to simulate.

steadyThreshold

the number of decimal places to which node values must be equivalent to be considered a steady state. This threshold must be passed for all nodes.

exogenousSupply

specifies if the value of a node (or nodes) is determined by an outside supply. In this case, the value of the node is supplied by the user and remains consistent throughout the course of the simulation. The default value for this argument is FALSE. If set to TRUE, there must exist an exogenousDef object in the model folder. This can be generated with the exogenousScreen function.

priorScreen

logical. Specifies if the function should collect modifier values generated from generated prior distributions. Default is set to FALSE.

robustnessTest

logical. Defaults to FALSE. Specifies if the nextStep function being used is part of a network robustness check.

genotypeBaseline

logical. Defaults to FALSE. Specified if only the first row of the genotype object should be used. If selected, the output object name will reflect that decision. If genotypeBaseline = TRUE and priorScreen = TRUE, a warning will be thrown to the user and the priorScreen argument will take precedence.

nodestartBaseline

logical. Defaults to FALSE. Specified if only the first row of the nodestart object should be used. If selected, the output object name will reflect that decision.

necStimThreshold

if the form of any necessary stimulant contains a threshold, the user must indicate said threshold and to which form the threshold will be applied. This indication will be achieved by passing a data.frame with the following columns: to, from, style, and threshold. to indicates the downstream node for the necessary stimulant. from the origin node, and style the name of the functional form for the necessary stimulant. The threshold column contains the values of the threshold parameter. Information only needs to be provided for cases where a threshold applies.

altTopologyName

default to NULL. If robustnessTest = TRUE, this argument allows the user to keep the generated alternate nextStep function with a specific name. If no name is provided, the alternate nextStep functions will be names nextStepAlt.R.

saveOutput

logical. Default set to TRUE. Indicates if the output of simulation screen should be automatically saved in the provided folder location upon completion.

combinatorial

logical. States if setupSims should simulate every combination of conditions that are being explored. If set to TRUE, every row from each set of explored conditions will be simulated against all rows from every other explored conditions (with the explored conditions relating to each data.frame that you have specified for setupSims to explore). If set to FALSE, it will be expected that all of the data.frames of conditions that have been passed to setupSims will have the same number of rows. In this case, each row will only be paired with the equivalent row of each data.frame. When this argument is set to FALSE, it is up to the user to make sure that all of the appropriate control conditions exist.

preventDrop

logical. Allows the user to override setupSims dropping any repeated conditions. This will also prevent reorganisation of conditions for clarity. Default is set to FALSE.

reduceSize

logical. Default set to FALSE. This argument is used to minimise the size of simulation output. If set to TRUE, only the final values of the simulation will be returned if the simulation reached stability. If the simulation was not able to reach stability, the full output of the simulation will be returned.

nCores

the number of cores to be used to run simulations. The default is set to 1, in which case simulations will be run sequentially on a singe core. If greater than 1, simulations will be run in parallel across the number of cores indicated. You can check the number of cores on your system with the function detectCores() from the parallel package. It is recommended that the maximum number of cores that you chose is at lease one less than the number returned by detectCores().