A function to simulate a network

Takes the output folder of the buildNetwork function and uses it to run a network simulation. The simulation will run until a steady state has been achieved unless otherwise specified. A steady state is considered to be reached when all node values are considered to be equivelant to the previous timestep for up to a threshold of decimal places.

Usage

simulateNetwork(
  folder,
  delay = 2,
  maxStep = NA,
  genotype = NA,
  startingValues = NA,
  steadyThreshold = 4,
  exogenousSupply = FALSE,
  necStimThreshold = NULL,
  robustnessTest = FALSE,
  altTopologyName = NULL,
  reduceSize = FALSE
)

Arguments

folder

the directory of the folder generated by the buildModel function.

delay

the amount of delay to apply for delayed action of nodes. The default is set to 2.

maxStep

the maximum number of steps that you want to simulate for. Will terminate simulation when steady state is reached, unless maxStep is reached first. If set to NA (the default), will simulate until stability is reached.

genotype

default set to NA. Allows the user to provide a different genotypeDef data.frame from the one specified in the provided file.

startingValues

default set to NA. Allows the user to provide a different nodestartDef data.frame from the one specified in the provided file.

steadyThreshold

the number of decimal places to which node values must be equivalent to be considered a steady state. This threshold must be passed for all nodes.

exogenousSupply

specifies if the value of a node (or nodes) is determined by an outside supply. In this case, the value of the node is supplied by the user and remains consistent throughout the course of the simulation. The default value for this argument is FALSE. If set to TRUE, there must exist an exogenousDef object in the model folder. This can be generated with the exogenousScreen function.

necStimThreshold

if the form of any necessary stimulant contains a threshold, the user must indicate said threshold and to which form the threshold will be applied. This indication will be achieved by passing a data.frame with the following columns: to, from, style, and threshold. to indicates the downstream node for the necessary stimulant. from the origin node, and style the name of the functional form for the necessary stimulant. The threshold column contains the values of the threshold parameter. Information only needs to be provided for cases where a threshold applies.

robustnessTest

logical. Defaults to FALSE. Specifies if the nextStep function being used is part of a network robustness check.

altTopologyName

default to NULL. If robustnessTest =TRUE, this argument allows the user to keep the generated alternate nextStep function with a specific name. If no name is provided, the alternate nextStep functions will be names nextStepAlt.R.

reduceSize

logical. Default set to FALSE. This argument is used to minimise the size of simulation output. If set to TRUE, only the final values of the simulation will be returned if the simulation reached stability. If the simulation was not able to reach stability, the full output of the simulation will be returned.

Note

this function can only be used if the output of the buildModel function was generated with the language argument set to "R".